airflow/requirements.txt
confluent-kafka
requests
python-dotenv
snowflake-connector-python